Fix and clarify threads and locking
Upon enabling `start_running` in the agent config, a deadlock was encountered
because the RobotConnector's `on_input_phase` would run before `self.pos` and
`self.yaw` were initialized, causing the thread that took the lock to die. I
wish I could get pylance to check that fields are definitelly initialized in the
__init__ call, but I don't see how right now.

Use the context API to manage locking/unlocking so that we don't have to be
careful about matching the calls and don't have to worry about a dead thread
causing deadlock again:
https://docs.python.org/3/library/threading.html#with-locks

Add a section to the readme that explains what threads are running in the app.

Name the Soar run thread so that it's readily recognizable during debugging.