Add reset button to ControllerGUI
This fully re-initializes the agent and scene. Refactor a bit for code re-use.
Add `echo=True` to all of the `ExecuteCommandLine` calls so that users watching
from the debugger can better see what's going on.

It would have been simple to delete the agent and re-create it, but that would
have required the user to re-connect the debugger. This approach is pretty
seamless.