Add a config option for printing SVS commands
The new config key is `log-svs-commands`.

For users that need to see all of the objects that are sent to SVS, this is
more accessible than adding print statements to the connector code.

Add documentation and add it to the example config. Add a `false` value to
John's config (not activating it, just letting him know that it's there).